mv load_tractogram_with_reference to io streamlines
