+++ scilpy/io/streamlines.py
@@ -23,3 +23,21 @@ def check_tracts_same_format(parser, tractogram_1, tractogram_2):
     if not ext_1 == ext_2:
         parser.error(
             'Input and output tractogram files must use the same format.')
+
+
+def load_tractogram_with_reference(parser, args, filepath,
+                                   bbox_check=True):
+    _, ext = os.path.splitext(filepath)
+    if ext == '.trk':
+        sft = load_tractogram(filepath, 'same',
+                              bbox_valid_check=bbox_check)
+    elif ext in ['.tck', '.fib', '.vtk', '.dpy']:
+        if args.reference is None:
+            parser.error('--reference is required for this file format '
+                         '{}.'.format(filepath))
+        sft = load_tractogram(filepath, args.reference,
+                              bbox_valid_check=bbox_check)
+    else:
+        parser.error('{} is an unsupported file format'.format(filepath))
+
+    return sft
