+++ scilpy/io/utils.py
@@ -19,24 +19,6 @@ def add_reference(parser):
                         'support (.nii or .nii.gz).')
 
 
-def load_tractogram_with_reference(parser, args, filepath,
-                                   bbox_check=True):
-    _, ext = os.path.splitext(filepath)
-    if ext == '.trk':
-        sft = load_tractogram(filepath, 'same',
-                              bbox_valid_check=bbox_check)
-    elif ext in ['.tck', '.fib', '.vtk', '.dpy']:
-        if args.reference is None:
-            parser.error('--reference is required for this file format '
-                         '{}.'.format(filepath))
-        sft = load_tractogram(filepath, args.reference,
-                              bbox_valid_check=bbox_check)
-    else:
-        parser.error('{} is an unsupported file format'.format(filepath))
-
-    return sft
-
-
 def add_overwrite_arg(parser):
     parser.add_argument(
         '-f', dest='overwrite', action='store_true',
