+++ scripts/scil_convert_tractogram.py
@@ -6,9 +6,9 @@ import os
 
 from dipy.io.streamline import save_tractogram
 
+from scilpy.io.streamlines import load_tractogram_with_reference
 from scilpy.io.utils import (add_overwrite_arg, add_reference,
-                             assert_inputs_exist, assert_outputs_exist,
-                             load_tractogram_with_reference)
+                             assert_inputs_exist, assert_outputs_exist)
 
 DESCRIPTION = """
 Conversion of '.tck', '.trk', '.fib', '.vtk' and 'dpy' files using updated file
