+++ scripts/scil_recognize_single_bundle.py
@@ -19,10 +19,10 @@ from dipy.tracking.streamline import transform_streamlines
 from nibabel.streamlines.array_sequence import ArraySequence
 import numpy as np
 
+from scilpy.io.streamlines import load_tractogram_with_reference
 from scilpy.io.utils import (add_overwrite_arg, add_reference, add_verbose_arg,
                              assert_inputs_exist,
-                             assert_outputs_exist,
-                             load_tractogram_with_reference)
+                             assert_outputs_exist)
 
 
 def _build_args_parser():
