+++ scripts/scil_remove_invalid_streamlines.py
@@ -5,9 +5,9 @@ import argparse
 
 from dipy.io.streamline import save_tractogram
 
+from scilpy.io.streamlines import load_tractogram_with_reference
 from scilpy.io.utils import (add_overwrite_arg, add_reference,
-                             assert_inputs_exist, assert_outputs_exist,
-                             load_tractogram_with_reference)
+                             assert_inputs_exist, assert_outputs_exist)
 
 DESCRIPTION = """
 Removal of streamlines that are out of the volume bounding box. In voxel space
